Use ICD(d) instead of f(d) for inv cumulative dist
We use f(x) throughout the book to mean many different things. In book 3
section 3, we use f(x) to mean the inverse cumulative distribution of x.
However, in section 4, we then switch to f(theta, phi) to mean a
function that we integrate over the surface of the unit sphere. This
appears in sphere_importance.cc in the same place that we had in
integrate_x_sq.cc, confusing the two functions to mean very different
things.

Dimitry Ishenko suggested changing to a more explicit name to indicate
the inverse cumulative distribution function, which helps make things
much more clear.

Resolves #1537